Fix warning

+++ Dev/Cpp/Effekseer/Effekseer.h
@@ -2431,7 +2431,7 @@ struct EffectBasicRenderParameter
 		int32_t ColorBlendType;
 		float BeginColor[4];
 		float EndColor[4];
-		int32_t Pow = 1;
+		float Pow = 1.0f;
 	} FalloffParam;
 
 	int32_t EmissiveScaling;
